WIPXX: factor out new polynomial struct to separate file
