+++ src/shared_math.rs
@@ -2,6 +2,7 @@ pub mod complex_number;
 pub mod fraction;
 pub mod low_degree_test;
 pub mod other;
+pub mod polynomial;
 pub mod polynomial_quotient_ring;
 pub mod prime_field_element;
 pub mod prime_field_polynomial;
